+++ index.html
@@ -13,7 +13,7 @@
 		<link rel="stylesheet" type="text/css" href="css/app.css">
 		<script src="node_modules/todomvc-common/base.js" type="application/javascript"></script>
 		<script src="node_modules/cydran/dist/cydran.min.js" type="application/javascript"></script>
-		<script src="app.js" type="application/javascript"></script>
+		<script src="app.js" type="module"></script>
 	</head>
 	<body>
 		<div id="appbody"></div>
